feat : 유저가 속한 특정 채팅방 조회기능 추가

+++ src/main/java/com/bubble/buubleforprofessor/chatroom/entity/ChatroomUser.java
@@ -38,10 +38,10 @@ public class ChatroomUser {
 @Embeddable
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 class ChatroomUserId implements Serializable {
-    private long chatroomId;
+    private int chatroomId;
     private UUID userId;
 
-    public ChatroomUserId(long chatroomId, UUID userId) {
+    public ChatroomUserId(int chatroomId, UUID userId) {
         this.chatroomId = chatroomId;
         this.userId = userId;
     }
